+++ packages/app-data/assets/global/android/icon-foreground.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:870dedca55ed85f48a426c4b5eae0e3d5fcabf255a116687af80693f21c751d3
-size 23932
+oid sha256:d9630898dac75f1178f5fd173af00b02f48cc2ce14062102f8fa1136e45d96d0
+size 21272
