+++ packages/app-data/assets/global/android/splash.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:9f6bb85531b30603b47f191b2b2bcbac53e42e97cd30039e8f64a46d54e2d28d
+size 82855
